feat: separate static and custom logo endpoints to prevent compilation blocking
- Refactor /api/logo to only handle static logo types via redirects
- Create /api/customLogo for heavy database/image processing functionality
- Update Logo and PoweredBy components to use /api/customLogo
- Resolves 3-second compilation blocking issue
